# api/openapi.yaml
openapi: 3.0.3
info:
  title: Waterflow API
  version: 1.1.0
  description: |
    声明式工作流编排引擎 REST API
    
    Waterflow 提供 YAML 驱动的分布式工作流编排能力，支持跨服务器任务执行、
    自动重试、完整的状态持久化和实时日志流。
    
    ## 架构更新 (ADR-0009)
    
    根据 [ADR-0009: 工作流定义与执行分离](../docs/adr/0009-workflow-definition-execution-separation.md)，
    API 已进行语义重构：
    
    - **Definition API** (`/v1/workflows`): 管理工作流定义（YAML + 元数据）
    - **Execution API** (`/v1/executions`): 管理执行实例
    - **Schedule API** (`/v1/workflows/{name}/schedules`): 挂载到定义下
    - **Webhook API** (`/v1/workflows/{name}/webhooks`): 挂载到定义下
    
    完整 API 清单详见 [api-inventory.md](../docs/api-inventory.md)。
    
    > **注意:** 当前 OpenAPI 规范尚未完全同步 ADR-0009 变更，将在后续版本更新。
    > 参考文档以 api-inventory.md 为准。
    
    ## 核心功能
    - 🎯 声明式 YAML DSL (类似 GitHub Actions)
    - 🔄 Event Sourcing 架构 - 完整执行历史追踪
    - 🌐 分布式 Agent 系统 - 跨服务器任务路由
    - 🔌 插件化节点系统 - 10+ 内置节点，支持自定义
    - 📊 实时日志流 - 支持 SSE 和 NDJSON 格式
    
    ## 认证
    当前版本支持两种认证方式:
    - **Bearer Token**: JWT 格式，用于用户认证
    - **API Key**: 在 HTTP Header 中传递 `X-API-Key`
    
    开发环境默认不需要认证。生产环境请配置 TLS 和认证。
  contact:
    name: Websoft9 Team
    url: https://github.com/websoft9/waterflow
    email: support@websoft9.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080
    description: 本地开发环境
  - url: https://api.waterflow.example.com
    description: 生产环境
...
security:
  - bearerAuth: []
  - apiKey: []
  - {} # Allow unauthenticated access for development
paths:
...
  /v1/agents:
    get:
      summary: 列出所有 Agent
      description: |
        返回所有已知 Task Queue 的 Agent 健康状态。
        通过 Temporal `DiscoverTaskQueues` 发现 Task Queue，再用 `DescribeTaskQueue` 查询实时 poller 数据。
      operationId: listAgents
      tags: [Agents]
      responses:
        '200':
          description: Agent 列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAgentsResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/agents/summary:
    get:
      summary: 获取 Agent 汇总统计
      description: 返回跨所有 Task Queue 的健康状态聚合数据，适用于监控仪表板。
      operationId: getAgentsSummary
      tags: [Agents]
      responses:
        '200':
          description: 汇总统计
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentsSummaryResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/agents/{name}:
    get:
      summary: 获取单个 Agent 状态
      description: 通过 Task Queue 名称查询指定 Agent 的实时 Temporal Worker 状态。
      operationId: getAgentStatus
      tags: [Agents]
      parameters:
        - name: name
          in: path
          required: true
          description: Agent 名称（即 Task Queue 名称）
          schema:
            type: string
            example: linux-amd64
      responses:
        '200':
          description: Agent 状态详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  # ==================== Task Queues ====================
  /v1/task-queues:
    get:
      summary: 列出所有 Task Queue
      description: 返回所有已知 Task Queue 及其实时 Worker 数量与健康状态（Temporal 原生数据）。
      operationId: listTaskQueues
      tags: [Task Queues]
      responses:
        '200':
          description: Task Queue 列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTaskQueuesResponse'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer Token 认证
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key 认证
...
  schemas:
...
    AgentResponse:
      type: object
      description: 单个 Agent（Task Queue）的健康状态
      properties:
        name:
          type: string
          example: linux-amd64
        pollers:
          type: integer
          description: Task Queue 上的 Worker 总数
          example: 3
        healthy_pollers:
          type: integer
          description: 30 秒内有活动的健康 Worker 数量
          example: 3
        task_backlog:
          type: integer
          format: int64
          description: 待处理任务数量
          example: 0
        status:
          type: string
          enum: [healthy, degraded, unavailable]
          description: |
            健康状态：
            - `healthy`：≥50% Worker 在 30 秒内活跃
            - `degraded`：>0 但 <50% Worker 健康
            - `unavailable`：无 Worker 或查询失败
          example: healthy
        last_update_time:
          type: string
          format: date-time
          example: "2026-03-06T10:30:00Z"
    ListAgentsResponse:
      type: object
      properties:
        agents:
          type: array
          items:
            $ref: '#/components/schemas/AgentResponse'
        total_count:
          type: integer
          example: 3
        timestamp:
          type: string
          format: date-time
          example: "2026-03-06T10:30:01Z"
    TaskQueueResponse:
      type: object
      description: Task Queue 实时状态（与 AgentResponse 结构相同，语义为队列视角）
      properties:
        name:
          type: string
          example: web-servers
        pollers:
          type: integer
          example: 2
        healthy_pollers:
          type: integer
          example: 1
        task_backlog:
          type: integer
          format: int64
          example: 5
        status:
          type: string
          enum: [healthy, degraded, unavailable]
          example: degraded
        last_update_time:
          type: string
          format: date-time
          example: "2026-03-06T10:29:00Z"
    ListTaskQueuesResponse:
      type: object
      properties:
        task_queues:
          type: array
          items:
            $ref: '#/components/schemas/TaskQueueResponse'
        total_count:
          type: integer
          example: 2
        timestamp:
          type: string
          format: date-time
          example: "2026-03-06T10:30:01Z"
    AgentsSummaryResponse:
      type: object
      description: 跨所有 Task Queue 的聚合健康统计
      properties:
        total_queues:
          type: integer
          example: 5
        healthy_queues:
          type: integer
          example: 3
        degraded_queues:
          type: integer
          example: 1
        unavailable_queues:
          type: integer
          example: 1
        total_pollers:
          type: integer
          example: 12
        healthy_pollers:
          type: integer
          example: 10
        timestamp:
          type: string
          format: date-time
          example: "2026-03-06T10:30:01Z"

    # ========== Workflow Schemas ==========
...
    ProblemDetails:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          description: 错误类型标识符
          example: "invalid_request"
        title:
          type: string
          description: 错误简短标题
          example: "Invalid Request"
        status:
          type: integer
          description: HTTP 状态码
          example: 400
        detail:
          type: string
          description: 错误详细说明
          example: "Missing required parameter: yaml"
        instance:
          type: string
          description: 请求路径
          example: "/v1/workflows"
        errors:
          type: array
          description: 详细错误列表（可选）
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: "invalid_request"
            title: "Invalid Request"
            status: 400
            detail: "Missing required parameter: yaml"
            instance: "/v1/workflows"
...
    InternalError:
      description: 服务器内部错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: "internal_error"
            title: "Internal Server Error"
            status: 500
            detail: "An unexpected error occurred"
            instance: "/v1/workflows"